requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
config==0.5.1
Django==4.0.2
django-rest-registration==0.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==9.0.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
